moving model processor into model package
